feat: home, commit list 연결

+++ presenter/src/main/java/com/mashup/twotoo/presenter/home/Home.kt
@@ -40,7 +40,7 @@ import org.orbitmvi.orbit.compose.collectSideEffect
 fun HomeRoute(
     homeViewModel: HomeViewModel,
     modifier: Modifier = Modifier,
-    navigateToHistory: () -> Unit = {},
+    navigateToHistory: (Int) -> Unit = {},
     navigateToCreateChallenge: () -> Unit = {},
 ) {
     val homeSideEffectHandler = rememberHomeSideEffectHandler(
@@ -109,7 +109,7 @@ fun HomeRoute(
 @Composable
 fun HomeScreen(
     modifier: Modifier = Modifier,
-    navigateToHistory: () -> Unit = {},
+    navigateToHistory: (Int) -> Unit = {},
     onBeeButtonClick: () -> Unit = {},
     onClickBeforeChallengeTextButton: (BeforeChallengeState) -> Unit = {},
     onCommit: () -> Unit = {},
